qemu_img_negative: Refine the case
Refine the negative case to create the images using the
API provided by avocado. Then it will be easy to run any
other backends, once avocado supports it.
