+++ qemu/tests/cfg/qemu_img_negative.cfg
@@ -1,31 +1,27 @@
 - qemu_img_negative:
     only qcow2
-    kill_vm = yes
-    start_vm = no
-    check_image = no
     variants:
         - rebase:
-            type = rebase_negative_test_on_iscsi
-            image_chain = "image1"
+            type = rebase_negative_test
+            images = "base sn"
+            image_name_base = "images/rebase_negative_base"
+            image_name_sn = "images/rebase_negative_sn"
+            force_create_image_base = yes
+            force_create_image_sn = yes
+            remove_image_base = yes
+            remove_image_sn = yes
             rebase_mode = safe
             image_format = qcow2
+            image_boot = no
+            kill_vm = yes
+            start_vm = no
+            negtive_test_sn = yes
             variants:
-                - new_to_base_lvm:
-                    storage_type = "iscsi"
-                    image_boot = no
-                    force_remove_image = yes
-                    vg_name = vgtest
-                    lv_size = 10G
-                    lv_name_list = "lv_base lv_sn1 lv_sn2"
-                    host_setup_flag = 2
-                    image_chain += " sn1"
-                    images += " sn1 sn2"
-                    image_name_image1 = "lv_base"
-                    image_name_sn1 = "lv_sn1"
-                    image_name_sn2 = "lv_sn2"
-                    rebase_list = "sn2 > image1"
-                    negtive_test_sn2 = yes
-                    #Please update portal_ip, initiator and target params based on your test env
-                    portal_ip = 127.0.0.1
-                    initiator = "iqn.2010-07.com.example:my_initiator"
-                    target = "iqn.2001-05.com.mytarget:0-0a0000-000000000-0000000000000000-my-target"
+                - nonexist_old_backing_file:
+                    rebase_list = "sn > base"
+                - nonexist_new_backing_file:
+                    images += " new"
+                    image_chain = "base sn"
+                    image_name_new = "images/rebase_negative_new"
+                    create_image_new = no
+                    rebase_list = "sn > new"
